refactor(backend): extract FritzBox VPN service into dedicated module
