+++ backend/app/services/vpn/__init__.py
@@ -9,11 +9,13 @@ Provides VPN management with:
 """
 
 from app.services.vpn.service import VPNService
+from app.services.vpn.fritzbox import FritzBoxVPNService
 from app.services.vpn.profiles import VPNService as VPNProfileService
 from app.services.vpn.encryption import VPNEncryption
 
 __all__ = [
     "VPNService",
+    "FritzBoxVPNService",
     "VPNProfileService",
     "VPNEncryption",
 ]
